Move PSP resource into domain package
