+++ internal/provider/psp/monitor_sort_test.go
@@ -1,4 +1,4 @@
-package provider
+package psp
 
 import (
 	"context"
