+++ internal/provider/psp/upgrader_test.go
@@ -1,4 +1,4 @@
-package provider
+package psp
 
 import (
 	"context"
